% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brim-main.R
\name{brim_ast}
\alias{brim_ast}
\title{Turn a Brim ZQL query into an abstract syntax tree}
\usage{
brim_ast(zql, host = brim_host())
}
\arguments{
\item{zql}{the ZQL query}

\item{host}{see \code{\link[=brim_host]{brim_host()}}}
}
\description{
Turn a Brim ZQL query into an abstract syntax tree
}
