% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brim-main.R
\name{brim_host}
\alias{brim_host}
\title{Retrieve the Brim host URL}
\usage{
brim_host()
}
\description{
Looks in the \code{BRIM_HOST} environment variable for the host
URL of the Brim instance to use for the API call. This
defaults to \verb{http://127.0.0.1:9867}.
}
