% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brim-main.R
\name{brim_spaces}
\alias{brim_spaces}
\title{Retrieve active Brim spaces from the specified Brim instance}
\usage{
brim_spaces(host = brim_host())
}
\arguments{
\item{host}{see \code{\link[=brim_host]{brim_host()}}}
}
\description{
Retrieve active Brim spaces from the specified Brim instance
}
