% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brimr-package.R
\docType{package}
\name{brimr}
\alias{brimr}
\alias{brimr-package}
\title{...}
\description{
A good description goes here otherwise CRAN checks fail.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.rud.is/hrbrmstr/brimr}
  \item Report bugs at \url{https://git.rud.is/hrbrmstr/brimr/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
