% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brim-main.R
\name{tidy_brim}
\alias{tidy_brim}
\title{Turn Brim/zqd search results into a data frame}
\usage{
tidy_brim(x)
}
\arguments{
\item{x}{Brim/zqd search results}
}
\description{
Turn Brim/zqd search results into a data frame
}
