% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brim-main.R
\name{brim_search}
\alias{brim_search}
\title{Post a ZQL query to the given Brim instance and retrieve processed results}
\usage{
brim_search(space_name, zql, host = brim_host())
}
\arguments{
\item{space_name}{name of the Brim space to use as the search data source}

\item{zql}{the ZQL query}

\item{host}{see \code{\link[=brim_host]{brim_host()}}}
}
\description{
Post a ZQL query to the given Brim instance and retrieve processed results
}
