% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brimr-package.R
\docType{package}
\name{brimr}
\alias{brimr}
\alias{brimr-package}
\title{Tools to Work with Brim and zqd}
\description{
Brim (\url{https://github.com/brimsec/brim}) enables efficient query
operations on large packet captures and log sources, such as Zeek. Tools
are provided to with with Brim components, including the Brim zqd query back-end.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.rud.is/hrbrmstr/brimr}
  \item Report bugs at \url{https://git.rud.is/hrbrmstr/brimr/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
